tests: Use session discarded counter to validate blocking test
Motivation
==========

On my development machine, the block test takes about 75s to complete,
with the majority of the time being used in the infinite blocking
test.

Solution
========

Using the discarded counter provided by `lttng list SESSION`, the test
no longer spends time reading and processing the ~1G of traces with
babeltrace2. This reduces the run time of the entire suite from 75s to
about 15s.

Known drawbacks
===============

There's no longer a validation that the trace files written to disk
are parseable by a CTF reader.

Change-Id: I0ccdef53ef80f1ffe5cfff970b92c3de4ba460ec
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>