Fix: consumerd: inaccurate math in discarded events aggregation
Observed issue
==============

When a discarded event counter overflow is encountered within a stream,
the math handling the overflow are wrong, both for 32-bit and 64-bit
consumer daemons.

The overflow should theoretically never be experienced on 64-bit
consumer daemons, except in scenarios where the discarded event count
goes backwards. This actually happens in stress-tests because the event
discarded counter is sampled in the check deliver which happens out of
order, but this is a separate issue.

On 32-bit, the overflow is expected to occur, because the ring buffer
tracks a 32-bit "unsigned long" value, which is then aggregated by
consumerd into a 64-bit counter.

Cause
=====

Incorrect maths in the handling of the counter overflow.

On 32-bit:

        if (discarded_events < stream->last_discarded_events) {
               /*
                * Overflow has occurred. We assume only one wrap-around
                * has occurred.
                */
               stream->chan->discarded_events += (1ULL << (CAA_BITS_PER_LONG - 1)) -
                       stream->last_discarded_events + discarded_events;
        }

The overflow detection is on 32-bit because discarded_events is sampled
from an "unsigned long" counter within lttng-ust, and
stream->last_discarded_events is the value of the prior discarded_event
for this stream.

(1ULL << (CAA_BITS_PER_LONG - 1)) is 2^31

2^31 + discarded_events - stream->last_discarded_events should offset
from 2^32 rather than 2^31, because it's handling a 2^32 overflow.

On 64-bit:

What we are aiming for is to update the discarded_events counter
with the new 64-bit value, but we end up calculating this incorrect
value derived from 2^63 instead:

        if (discarded_events < stream->last_discarded_events) {
               /*
                * Overflow has occurred. We assume only one wrap-around
                * has occurred.
                */
               stream->chan->discarded_events += (1ULL << (CAA_BITS_PER_LONG - 1)) -
                       stream->last_discarded_events + discarded_events;
        }

1ULL << (CAA_BITS_PER_LONG - 1) is 2^63.

2^63 + discarded_events - stream->last_discarded_events offsets from
2^63 rather than 0.

Solution
========

Given that rollback of the per-stream event discarded counter is not
expected on 64-bit, print an error when this situation is detected.
Do not update the channel counter if overflow is detected on 64-bit,
so the situation can be recovered on the next aggregation if it was
caused by a spurious race condition.

On 32-bit, when a 32-bit overflow is detected, add (1ULL << 32) to the
delta.

Known drawbacks
===============

Those stress-test scenarios will emit the new error message when a
64-bit consumer daemon encounters out-of-order ring buffer check
deliver, which causes the discarded event counter to go backward for
consecutive subbuffers within a stream.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I1a0fca2dcf6e3fe5ec37509838515545b5ee9062