Tests: Add flags to test lttng userspace only
The current tests will run both userspace and kernel testing.
Some of use cases only use lttng for one kind of tracing on an embedded device (e.g. userspace)
If the lttng modules(.ko files) is not present during the test, it would end up with lots
of failing.

This commit separate the tests by allowing a flag to be passed and marked the kernel related
tests as SKIP instead of FAIL if the modules is not presented.

e.g. before running the tests with make, add a flag to environment like "export SKIP_KERNEL_TEST=1"
the test cases would skip kernel test.

Tested with this commit under a local setup, the result as below:

====================================================
   lttng-tools 2.13.11: tests/unit/test-suite.log
====================================================

 TOTAL: 1032
 PASS:  1032
 SKIP:  0
 XFAIL: 0
 FAIL:  0
 XPASS: 0
 ERROR: 0

.. contents:: :depth: 2

==========================================================
   lttng-tools 2.13.11: tests/regression/test-suite.log
==========================================================

 TOTAL: 11091
 PASS:  7243
 SKIP:  3848
 XFAIL: 0
 FAIL:  0
 XPASS: 0
 ERROR: 0

.. contents:: :depth: 2

PASS: tools/filtering/test_valid_filter
=======================================

Signed-off-by: Xiangyu Chen <xiangyu.chen@aol.com>