Docs: Add the example commit from CONTRIBUTING as a template
Observed issue
==============

I found myself referring to the contributing guide each time a made a
new commit to ensure that I had the appropriate style and sections.

Solution
========

A stub template `.commit_template` has been added and the instructions
in `CONTRIBUTING.md` indicate that the local checkout that be
configured to use it as a template for new commits which are made.

Known drawbacks
===============

None.

Change-Id: Idfacd4d726657cb57f193f0a3375a840d8a9c746
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>