Tests: lttngtest: confusing comment regarding supported python versions
The comment in _get_time_ns() caused me to do a double-take since the
function checks for Python > 3.3 and mentions Python 3.8 for unrelated
reasons.

I am clarifying the comments a bit to explain the reason for the two
versions being mentionned.

Also, the version check is changed to match 3.3 (although we don't
support that version) since that is when time.monotonic was introduced.

A type annotation is also added to clarify the function's intended usage
(i.e., it will not return fractional nanoseconds).

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I9f4fd7cf1d0673e80f6bce9e3cf86b9697fe3f91